fix: change stock from bigdecimal to long in various classes that refer to product entity

+++ src/main/java/com/harriet/shopiify/product/vo/ProductVO.java
@@ -37,7 +37,7 @@ public class ProductVO {
 
     @NotNull
     @Min(value = 0L, message = "The value must be positive")
-    private BigInteger stock;
+    private Long stock;
 
     @NotNull
     @Min(value = 0, message = "The value must be positive")
